requirements.txt
click==8.1.8
Jinja2==3.1.5
MarkupSafe==3.0.2
PyYAML==6.0.2
